Commit Final Proyecto Terminado

+++ src/main/java/com/alura/foroalura/controller/CursoController.java
@@ -15,6 +15,14 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+/**
+ * Esta clase es un controlador de Spring que maneja las solicitudes
+ * relacionadas con los cursos en la aplicación. Proporciona endpoints para
+ * crear, obtener, actualizar y eliminar cursos.
+ * 
+ * @author Saúl Wade León
+ * @version 1.3
+ */
 @RestController
 @RequestMapping("/cursos")
 @SecurityRequirement(name = "bearer-key")
@@ -31,6 +39,13 @@ public class CursoController {
     @PostMapping
     @Transactional
     @Operation(summary = "Crear un nuevo curso")
+    /**
+     * Este método maneja las solicitudes POST en la ruta "/cursos" y se encarga de
+     * crear un nuevo curso.
+     * 
+     * @param curso
+     * @return
+     */
     public ResponseEntity<Curso> crearCurso(@Valid @RequestBody Curso curso) {
         Curso nuevoCurso = cursoService.crearCurso(curso);
         return new ResponseEntity<>(nuevoCurso, HttpStatus.CREATED);
@@ -39,6 +54,12 @@ public class CursoController {
     // Endpoint para obtener todos los cursos
     @GetMapping
     @Operation(summary = "Obtener el listado de todos los cursos")
+    /**
+     * Este método maneja las solicitudes GET en la ruta "/cursos" y se encarga de
+     * obtener todos los cursos disponibles.
+     * 
+     * @return
+     */
     public ResponseEntity<List<Curso>> obtenerTodosLosCursos() {
         List<Curso> cursos = cursoService.obtenerTodosLosCursos();
         return new ResponseEntity<>(cursos, HttpStatus.OK);
@@ -47,6 +68,13 @@ public class CursoController {
     // Endpoint para obtener un curso por su ID
     @GetMapping("/{id}")
     @Operation(summary = "Obtener un curso por su ID")
+    /**
+     * Este método maneja las solicitudes GET en la ruta "/cursos/{id}" y se encarga
+     * de obtener un curso por su ID.
+     * 
+     * @param id
+     * @return
+     */
     public ResponseEntity<Curso> obtenerCursoPorId(@PathVariable Long id) {
         Curso curso = cursoService.obtenerCursoPorId(id);
         if (curso != null) {
@@ -60,6 +88,14 @@ public class CursoController {
     @PutMapping("/{id}")
     @Transactional
     @Operation(summary = "Actualizar un curso por su ID")
+    /**
+     * Este método maneja las solicitudes PUT en la ruta "/cursos/{id}" y se encarga
+     * de actualizar un curso por su ID.
+     * 
+     * @param id
+     * @param cursoActualizado
+     * @return
+     */
     public ResponseEntity<Curso> actualizarCurso(@PathVariable Long id, @Valid @RequestBody Curso cursoActualizado) {
         Curso curso = cursoService.actualizarCurso(id, cursoActualizado);
         if (curso != null) {
@@ -72,8 +108,19 @@ public class CursoController {
     // Endpoint para eliminar un curso por su ID
     @DeleteMapping("/{id}")
     @Operation(summary = "Eliminar un curso por su ID")
-    public ResponseEntity<Void> eliminarCurso(@PathVariable Long id) {
-        cursoService.eliminarCurso(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    /**
+     * Este método maneja las solicitudes DELETE en la ruta "/cursos/{id}" y se
+     * encarga de eliminar un curso por su ID.
+     * 
+     * @param id
+     * @return
+     */
+    public ResponseEntity<String> eliminarCurso(@PathVariable Long id) {
+        if (cursoService.existsById(id)) {
+            cursoService.eliminarCurso(id);
+            return new ResponseEntity<>("El curso fue eliminado exitosamente", HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>("El curso que intenta eliminar no existe", HttpStatus.NOT_FOUND);
+        }
     }
 }
