+++ src/main/java/com/alura/foroalura/controller/UsuarioController.java
@@ -24,6 +24,15 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.stream.Collectors;
 
+/**
+ * Esta clase es un controlador de Spring que maneja las solicitudes
+ * relacionadas con los usuarios en la aplicación. Proporciona endpoints para
+ * crear, obtener, actualizar, eliminar y desactivar usuarios, así como para
+ * consultar todos los usuarios disponibles.
+ * 
+ * @author Saúl Wade León
+ * @version 1.3
+ */
 @RestController
 @RequestMapping("/usuarios")
 @SecurityRequirement(name = "bearer-key")
@@ -48,6 +57,13 @@ public class UsuarioController {
     @PostMapping
     @Transactional
     @Operation(summary = "Crear un nuevo usuario")
+    /**
+     * Este método maneja las solicitudes POST en la ruta "/usuarios"
+     * y se encarga de crear un nuevo usuario.
+     * 
+     * @param datosRegistroUsuario
+     * @return
+     */
     public ResponseEntity<String> crearUsuario(@RequestBody @Valid DatosRegistroUsuario datosRegistroUsuario) {
         Role role = roleRepository.findByNombreRole(datosRegistroUsuario.role()); // Busca el rol por nombre
 
@@ -72,52 +88,15 @@ public class UsuarioController {
                 HttpStatus.CREATED);
     }
 
-    /*
-     * public ResponseEntity<Usuario> crearUsuario(@RequestBody @Valid
-     * DatosRegistroUsuario datosRegistroUsuario) {
-     * Role role = roleRepository.findByNombreRole(datosRegistroUsuario.role()); //
-     * Busca el rol por nombre
-     * 
-     * // Verificar si el nombre de usuario ya existe
-     * if (usuarioService.existeUsuarioPorUsername(datosRegistroUsuario.username()))
-     * {
-     * throw new
-     * UsernameAlreadyExistsException("El nombre de usuario ya está en uso.");
-     * }
-     * 
-     * if (role == null) {
-     * // Manejar el caso donde el rol no existe
-     * System.out.println("El rol especificado no existe");
-     * }
-     * 
-     * // Verificar si el rol existe en el catálogo de roles
-     * if (!usuarioService.existeRole(datosRegistroUsuario.role())) {
-     * throw new
-     * InvalidRoleException("El role de usuario proporcionado no es válido.");
-     * }
-     * 
-     * // Crear un nuevo usuario con los datos proporcionados
-     * Usuario nuevoUsuario = new Usuario(datosRegistroUsuario, role);
-     * 
-     * /*
-     * Usuario nuevoUsuario = new Usuario(
-     * null,
-     * datosRegistroUsuario.nombre(),
-     * datosRegistroUsuario.apellido(),
-     * datosRegistroUsuario.email(),
-     * datosRegistroUsuario.username(),
-     * datosRegistroUsuario.password(),
-     * role.getId());
-     * 
-     * 
-     * Usuario usuarioCreado = usuarioService.crearUsuario(nuevoUsuario);
-     * return new ResponseEntity<>(usuarioCreado, HttpStatus.CREATED);
-     * }
-     */
-
     // Endpoint para obtener todos los usuarios y mostrar los datos de un usuario
     @GetMapping
     @Operation(summary = "Obtener la lista de todos los usuarios")
+    /**
+     * Este método maneja las solicitudes GET en la ruta "/usuarios" y se encarga de
+     * obtener todos los usuarios disponibles.
+     * 
+     * @return
+     */
     public ResponseEntity<List<DatosDelUsuario>> obtenerTodosLosUsuarios() {
         List<Usuario> usuarios = usuarioService.obtenerTodosLosUsuarios();
         List<DatosDelUsuario> usuariosInfo = usuarios.stream()
@@ -130,6 +109,13 @@ public class UsuarioController {
     // Endpoint para obtener un usuario por su ID y mostrar sus datos
     @GetMapping("/{id}")
     @Operation(summary = "Obtener un usuario por su ID")
+    /**
+     * Este método maneja las solicitudes GET en la ruta "/usuarios/{id}" y se
+     * encarga de obtener los datos de un usuario por su ID.
+     * 
+     * @param id
+     * @return
+     */
     public ResponseEntity<DatosDelUsuario> obtenerUsuarioPorId(@PathVariable Long id) {
         Usuario usuario = usuarioService.obtenerUsuarioPorId(id);
         if (usuario != null) {
@@ -144,7 +130,13 @@ public class UsuarioController {
     @PutMapping("/{id}")
     @Transactional
     @Operation(summary = "Actualizar un usuario por su ID")
-
+    /**
+     * Este método maneja las solicitudes PUT en la ruta "/usuarios/{id}" y se
+     * encarga de actualizar los datos de un usuario por su ID.
+     * 
+     * @param id
+     * @return
+     */
     public ResponseEntity<String> actualizarUsuario(@PathVariable Long id,
             @RequestBody DatosActualizacionUsuario datosActualizacionUsuario) {
         Usuario usuario = usuarioService.obtenerUsuarioPorId(id);
@@ -185,14 +177,33 @@ public class UsuarioController {
     @DeleteMapping("/{id}")
     @Transactional
     @Operation(summary = "Eliminar un usuario por su ID")
-    public ResponseEntity<Void> eliminarUsuario(@PathVariable Long id) {
-        usuarioService.eliminarUsuario(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    /**
+     * Este método maneja las solicitudes DELETE en la ruta "/usuarios/{id}" y
+     * se encarga de eliminar un usuario por su ID.
+     * 
+     * @param id
+     * @return
+     */
+    public ResponseEntity<String> eliminarUsuario(@PathVariable Long id) {
+        if (usuarioService.existsById(id)) {
+            usuarioService.eliminarUsuario(id);
+            return new ResponseEntity<>("El usuario fue eliminado exitosamente", HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>("El usuario que intenta eliminar no existe", HttpStatus.NOT_FOUND);
+        }
     }
 
     // Endpoint para desactivar un usuario por su ID (borrado lógico)
     @PatchMapping("/{id}/desactivar")
     @Operation(summary = "Desactivar un usuario por su ID")
+    /**
+     * Este método maneja las solicitudes PATCH en la ruta
+     * "/usuarios/{id}/desactivar"
+     * y se encarga de desactivar un usuario por su ID.
+     * 
+     * @param id
+     * @return
+     */
     public ResponseEntity<Void> desactivarUsuario(@PathVariable Long id) {
         usuarioService.desactivarUsuario(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
